+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/service/TextItemService.kt
@@ -24,7 +24,7 @@ class TextItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
     }
 
     override fun createPreClipItem(
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         identifier: String,
         transferable: Transferable,
@@ -39,7 +39,7 @@ class TextItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
 
     override fun doLoadRepresentation(
         transferData: Any,
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         dataFlavor: DataFlavor,
         dataFlavorMap: Map<String, List<DataFlavor>>,
