+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/service/UrlItemService.kt
@@ -21,7 +21,7 @@ class UrlItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
     }
 
     override fun createPreClipItem(
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         identifier: String,
         transferable: Transferable,
@@ -36,7 +36,7 @@ class UrlItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
 
     override fun doLoadRepresentation(
         transferData: Any,
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         dataFlavor: DataFlavor,
         dataFlavorMap: Map<String, List<DataFlavor>>,
